pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "icedrive-blob"
version = "0.0.1"
authors = [
  { name="Marcelo Chinarro Cabrero", email="marcelo.chinarro@alu.uclm.es" },
]
description = "Implementation of IceDrive Blob service"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"
dependencies = [
    "zeroc-ice>=3.7",
]

[project.optional-dependencies]
tests = ["pytest"]
linters = ["black", "ruff", "pydocstyle"]

[project.urls]
"Homepage" = "https://github.com/Sikander035/SSDD_MarceloChinarroCabrero_BlobService"

[project.scripts]
icedrive-blob = "icedrive_blob.command_line_handler:server"
icedrive-client = "icedrive_blob.command_line_handler:client"
